refactor(#278): split routes/workspaces.ts into domain modules
- Extract review/comment/todo endpoints into routes/workspace-reviews.ts
- Move withWorkspaceOperationLock to utils/operation-lock.ts
- Move normalizeAdditionalDirectories to utils/paths.ts
- Move autoCommitWorktree to services/worktree.ts
- Main workspaces.ts now assembles sub-routers, keeping same public API
